chore: move populate.go to utils
